nova release

+++ database/migrations/2019_02_12_225526_create_foto_animals_table.php
@@ -15,7 +15,7 @@ class CreateFotoAnimalsTable extends Migration
     {
         Schema::create('foto_animals', function (Blueprint $table) {
             $table->increments('id_foto_animals');
-            $table->integer('id_animal')->unsigned();
+            $table->bigInteger('id_animal')->unsigned();
             $table->string('foto_animal', 1200);
 
             $table->foreign('id_animal')->references('id_animal')->on('animals')->onDelete('cascade');
